+++ app/js/all.js
@@ -1,3 +1,2 @@
-require('./ui');
-
+export {initializeUI} from './ui';
 export {TextInput} from './components/Forms';
