+++ lib/ui/forms.js
@@ -1,385 +1,3 @@
-(function() {
-  'use strict';
-
-  var globals = typeof global === 'undefined' ? self : global;
-  if (typeof globals.require === 'function') return;
-
-  var modules = {};
-  var cache = {};
-  var aliases = {};
-  var has = {}.hasOwnProperty;
-
-  var expRe = /^\.\.?(\/|$)/;
-  var expand = function(root, name) {
-    var results = [], part;
-    var parts = (expRe.test(name) ? root + '/' + name : name).split('/');
-    for (var i = 0, length = parts.length; i < length; i++) {
-      part = parts[i];
-      if (part === '..') {
-        results.pop();
-      } else if (part !== '.' && part !== '') {
-        results.push(part);
-      }
-    }
-    return results.join('/');
-  };
-
-  var dirname = function(path) {
-    return path.split('/').slice(0, -1).join('/');
-  };
-
-  var localRequire = function(path) {
-    return function expanded(name) {
-      var absolute = expand(dirname(path), name);
-      return globals.require(absolute, path);
-    };
-  };
-
-  var initModule = function(name, definition) {
-    var hot = hmr && hmr.createHot(name);
-    var module = {id: name, exports: {}, hot: hot};
-    cache[name] = module;
-    definition(module.exports, localRequire(name), module);
-    return module.exports;
-  };
-
-  var expandAlias = function(name) {
-    return aliases[name] ? expandAlias(aliases[name]) : name;
-  };
-
-  var _resolve = function(name, dep) {
-    return expandAlias(expand(dirname(name), dep));
-  };
-
-  var require = function(name, loaderPath) {
-    if (loaderPath == null) loaderPath = '/';
-    var path = expandAlias(name);
-
-    if (has.call(cache, path)) return cache[path].exports;
-    if (has.call(modules, path)) return initModule(path, modules[path]);
-
-    throw new Error("Cannot find module '" + name + "' from '" + loaderPath + "'");
-  };
-
-  require.alias = function(from, to) {
-    aliases[to] = from;
-  };
-
-  var extRe = /\.[^.\/]+$/;
-  var indexRe = /\/index(\.[^\/]+)?$/;
-  var addExtensions = function(bundle) {
-    if (extRe.test(bundle)) {
-      var alias = bundle.replace(extRe, '');
-      if (!has.call(aliases, alias) || aliases[alias].replace(extRe, '') === alias + '/index') {
-        aliases[alias] = bundle;
-      }
-    }
-
-    if (indexRe.test(bundle)) {
-      var iAlias = bundle.replace(indexRe, '');
-      if (!has.call(aliases, iAlias)) {
-        aliases[iAlias] = bundle;
-      }
-    }
-  };
-
-  require.register = require.define = function(bundle, fn) {
-    if (bundle && typeof bundle === 'object') {
-      for (var key in bundle) {
-        if (has.call(bundle, key)) {
-          require.register(key, bundle[key]);
-        }
-      }
-    } else {
-      modules[bundle] = fn;
-      delete cache[bundle];
-      addExtensions(bundle);
-    }
-  };
-
-  require.list = function() {
-    var list = [];
-    for (var item in modules) {
-      if (has.call(modules, item)) {
-        list.push(item);
-      }
-    }
-    return list;
-  };
-
-  var hmr = globals._hmr && new globals._hmr(_resolve, require, modules, cache);
-  require._cache = cache;
-  require.hmr = hmr && hmr.wrap;
-  require.brunch = true;
-  globals.require = require;
-})();
-
-(function() {
-var global = typeof window === 'undefined' ? this : window;
-var process;
-var __makeRelativeRequire = function(require, mappings, pref) {
-  var none = {};
-  var tryReq = function(name, pref) {
-    var val;
-    try {
-      val = require(pref + '/node_modules/' + name);
-      return val;
-    } catch (e) {
-      if (e.toString().indexOf('Cannot find module') === -1) {
-        throw e;
-      }
-
-      if (pref.indexOf('node_modules') !== -1) {
-        var s = pref.split('/');
-        var i = s.lastIndexOf('node_modules');
-        var newPref = s.slice(0, i).join('/');
-        return tryReq(name, newPref);
-      }
-    }
-    return none;
-  };
-  return function(name) {
-    if (name in mappings) name = mappings[name];
-    if (!name) return;
-    if (name[0] !== '.' && pref) {
-      var val = tryReq(name, pref);
-      if (val !== none) return val;
-    }
-    return require(name);
-  }
-};
-require.register("js/App.js", function(exports, require, module) {
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _reactRouterDom = require('react-router-dom');
-
-var _Home = require('./containers/Home');
-
-var _Home2 = _interopRequireDefault(_Home);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
-// import Components from './containers/Components';
-// import Typography from './containers/Typography';
-
-var App = function (_Component) {
-  _inherits(App, _Component);
-
-  function App() {
-    _classCallCheck(this, App);
-
-    return _possibleConstructorReturn(this, (App.__proto__ || Object.getPrototypeOf(App)).apply(this, arguments));
-  }
-
-  _createClass(App, [{
-    key: 'render',
-    value: function render() {
-      return _react2.default.createElement(
-        _reactRouterDom.HashRouter,
-        null,
-        _react2.default.createElement(
-          'div',
-          { className: 'flex-wrapper' },
-          _react2.default.createElement(_reactRouterDom.Route, { exact: true, path: '/', component: _Home2.default })
-        )
-      );
-    }
-  }]);
-
-  return App;
-}(_react.Component);
-
-exports.default = App;
-
-});
-
-require.register("js/all.js", function(exports, require, module) {
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _ui = require('./ui');
-
-Object.defineProperty(exports, 'initializeUI', {
-  enumerable: true,
-  get: function get() {
-    return _ui.initializeUI;
-  }
-});
-
-var _Forms = require('./components/Forms');
-
-Object.defineProperty(exports, 'TextInput', {
-  enumerable: true,
-  get: function get() {
-    return _Forms.TextInput;
-  }
-});
-
-});
-
-require.register("js/components/Forms.js", function(exports, require, module) {
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.TextInput = undefined;
-
-var _react = require("react");
-
-var _react2 = _interopRequireDefault(_react);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function noop() {}
-
-var TextInput = exports.TextInput = function TextInput(props) {
-  return _react2.default.createElement(
-    "div",
-    { className: "input-field" },
-    _react2.default.createElement("input", { id: "first_name", type: "text", onChange: props.onChange || noop,
-      className: "validate", defaultValue: props.value }),
-    _react2.default.createElement(
-      "label",
-      { htmlFor: "first_name" },
-      props.label
-    )
-  );
-};
-
-});
-
-require.register("js/containers/Home.js", function(exports, require, module) {
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _all = require('../all');
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var Home = function Home(props) {
-  return _react2.default.createElement(
-    'div',
-    { className: 'primary-container home-container flex-wrapper row with-sidenav' },
-    _react2.default.createElement(
-      'div',
-      { className: 'flex-column nav-column' },
-      _react2.default.createElement(
-        'div',
-        { className: 'side-nav fixed' },
-        _react2.default.createElement('ul', null)
-      )
-    ),
-    _react2.default.createElement(
-      'div',
-      { className: 'flex-column main-content-container' },
-      _react2.default.createElement(_all.TextInput, { label: 'Example Text Field' })
-    )
-  );
-};
-
-exports.default = Home;
-
-});
-
-require.register("js/index.js", function(exports, require, module) {
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _reactDom = require('react-dom');
-
-var _redux = require('redux');
-
-var _lodash = require('lodash');
-
-var _isdev = require('isdev');
-
-var _isdev2 = _interopRequireDefault(_isdev);
-
-var _App = require('./App');
-
-var _App2 = _interopRequireDefault(_App);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-// import reducers from './reducers';
-
-var store = {};
-//
-
-// import thunk from 'redux-thunk';
-
-
-var Root = function Root(props) {
-  return _react2.default.createElement(_App2.default, { store: store });
-};
-
-exports.default = Root;
-
-});
-
-require.register("js/ui.js", function(exports, require, module) {
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.initializeUI = undefined;
-
-var _jquery = require('jquery');
-
-var _jquery2 = _interopRequireDefault(_jquery);
-
-var _hammerjs = require('hammerjs');
-
-var _hammerjs2 = _interopRequireDefault(_hammerjs);
-
-var _forms = require('./ui/forms');
-
-var _forms2 = _interopRequireDefault(_forms);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var initializeUI = exports.initializeUI = function initializeUI() {
-  (0, _forms2.default)(_jquery2.default);
-};
-
-});
-
-require.register("js/ui/forms.js", function(exports, require, module) {
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
@@ -1184,13 +802,3 @@ var formsUI = function formsUI($) {
 };
 
 exports.default = formsUI;
-
-});
-
-require.alias("buffer/index.js", "buffer");
-require.alias("process/browser.js", "process");process = require('process');require.register("___globals___", function(exports, require, module) {
-  
-});})();require('___globals___');
-
-
-//# sourceMappingURL=app.js.map
